create service commented

+++ src/request-sessions/request-sessions.service.ts
@@ -5,6 +5,7 @@ import { PrismaClient } from 'generated/prisma';
 import { PrismaException }          from '@config/prisma-catch';
 import { CreateRequestSessionDto }  from '@request-sessions/dto/create-request-session.dto';
 import { UpdateRequestSessionDto }  from '@request-sessions/dto/update-request-session.dto';
+import { UpdateSessionDayModulesDto } from './dto/update-session-day-modules.dto';
 
 
 @Injectable()
@@ -14,40 +15,72 @@ export class RequestSessionsService extends PrismaClient implements OnModuleInit
 		this.$connect();
 	}
 
+    #selectRequestSession = {
+        id              : true,
+        session         : true,
+        spaceId         : true,
+        isEnglish       : true,
+        isConsecutive   : true,
+        inAfternoon     : true,
+        description     : true,
+        spaceType       : true,
+        building        : true,
+        // requestId       : true,
+        professor       : {
+            select : {
+                id: true,
+                name: true,
+            }
+        },
+        createdAt       : true,
+        updatedAt       : true,
+        spaceSize       : {
+            select : {
+                id      : true,
+                detail  : true,
+            }
+        },
+        sessionDayModules: {
+            select : {
+                dayModuleId: true
+            }
+        }
+    }
 
-	async create( requestId: string, createRequestSessionDto: CreateRequestSessionDto ) {
-		try {
-			const { dayModulesId, ...data } = createRequestSessionDto;
 
-			const requestSession = await this.requestSession.create({
-				data : {
-					...data,
-					requestId
-				},
-			});
+	// async create( requestId: string, createRequestSessionDto: CreateRequestSessionDto ) {
+	// 	try {
+	// 		const { dayModulesId, ...data } = createRequestSessionDto;
 
-			await this.sessionDayModule.createMany({
-				data : dayModulesId.map(( dayModuleId ) => ({
-					dayModuleId,
-					requestSessionId : requestSession.id
-				})),
-			});
+	// 		const requestSession = await this.requestSession.create({
+	// 			data : {
+	// 				...data,
+	// 				requestId
+	// 			},
+	// 		});
 
-			// const requestSessionData = this.#requestSessionMap( requestSession );
+	// 		await this.sessionDayModule.createMany({
+	// 			data : dayModulesId.map(( dayModuleId ) => ({
+	// 				dayModuleId,
+	// 				requestSessionId : requestSession.id
+	// 			})),
+	// 		});
 
-			// this.sseService.emitEvent({
-			//     message : requestSessionData,
-			//     action  : EnumAction.CREATE,
-			//     type    : Type.REQUEST_SESSION,
-			//     origin
-			// });
+	// 		// const requestSessionData = this.#requestSessionMap( requestSession );
 
-			// return requestSessionData;
-			return requestSession;
-		} catch ( error ) {
-			throw PrismaException.catch( error, 'Failed to create request session' );
-		}
-	}
+	// 		// this.sseService.emitEvent({
+	// 		//     message : requestSessionData,
+	// 		//     action  : EnumAction.CREATE,
+	// 		//     type    : Type.REQUEST_SESSION,
+	// 		//     origin
+	// 		// });
+
+	// 		// return requestSessionData;
+	// 		return requestSession;
+	// 	} catch ( error ) {
+	// 		throw PrismaException.catch( error, 'Failed to create request session' );
+	// 	}
+	// }
 
 
     #requestSessionMap = ( requestSession ) => ({
@@ -62,42 +95,11 @@ export class RequestSessionsService extends PrismaClient implements OnModuleInit
 		try {
 			const requestSessions = await this.requestSession.findMany({
 				where : { requestId },
-                select : {
-                    id              : true,
-                    session         : true,
-                    spaceId         : true,
-                    isEnglish       : true,
-                    isConsecutive   : true,
-                    inAfternoon     : true,
-                    description     : true,
-                    spaceType       : true,
-                    building        : true,
-                    professor       : {
-                        select : {
-                            id: true,
-                            name: true,
-                        }
-                    },
-                    createdAt       : true,
-                    updatedAt       : true,
-                    spaceSize       : {
-                        select : {
-                            id      : true,
-                            detail  : true,
-                        }
-                    },
-                    sessionDayModules: {
-                        select : {
-                            dayModuleId: true
-                        }
-                    }
-                }
+                select : this.#selectRequestSession
 			});
 
 			return requestSessions.map(( requestSession ) => this.#requestSessionMap( requestSession ));
-			// return requestSessions;
 		} catch ( error ) {
-			console.log('🚀 ~ file: request-sessions.service.ts:136 ~ error:', error)
 			throw PrismaException.catch( error, 'Failed to find request sessions' );
 		}
 	}
@@ -107,15 +109,14 @@ export class RequestSessionsService extends PrismaClient implements OnModuleInit
 		try {
 			const requestSession = await this.requestSession.findUnique({
 				where : { id },
-				// select  : this.#selectRequestSession
+				select  : this.#selectRequestSession
 			});
 
 			if ( !requestSession ) {
 				throw new NotFoundException( 'Request session not found' );
 			}
 
-			// const requestSessionData = this.#requestSessionMap( requestSession );
-			return requestSession;
+			return this.#requestSessionMap( requestSession );
 		} catch ( error ) {
 			throw PrismaException.catch( error, 'Failed to find request session' );
 		}
@@ -124,41 +125,74 @@ export class RequestSessionsService extends PrismaClient implements OnModuleInit
 
 	async update( id: string, updateRequestSessionDto: UpdateRequestSessionDto ) {
 		try {
-			const { dayModulesId, ...data } = updateRequestSessionDto;
-
 			const requestSession = await this.requestSession.update({
-				where : { id },
-				data
+                select  : this.#selectRequestSession,
+				where   : { id },
+                data    : updateRequestSessionDto,
 			});
 
-			const dayModulesByRequestSession = await this.sessionDayModule.findMany({
-				where : { requestSessionId: id },
-			});
 
-			const currentDayModuleIds   = dayModulesByRequestSession.map(( dm ) => dm.dayModuleId ).sort();
-			const newDayModuleIds       = ( dayModulesId || [] ).sort();
-			const areEqual              = currentDayModuleIds.length === newDayModuleIds.length &&
-				currentDayModuleIds.every(( id, index ) => id === newDayModuleIds[index] );
+            // await this.request.update({
+            //     where : { id: requestSession.requestId },
+            //     data  : { status: 'APPROVED' },
+            // });
 
-			if ( !areEqual ) {
-				await this.sessionDayModule.deleteMany({
-					where : { requestSessionId: id },
+            return this.#requestSessionMap( requestSession )
+		} catch ( error ) {
+			throw PrismaException.catch( error, 'Failed to update request session' );
+		}
+	}
+
+	// Update session day modules for multiple request sessions
+	async updateSessionDayModules( updateSessionDayModulesDto: UpdateSessionDayModulesDto[] ) {
+		try {
+			for ( const dto of updateSessionDayModulesDto ) {
+				const { requestSessionId, dayModulesId } = dto;
+
+				const currentSessionDayModules = await this.sessionDayModule.findMany({
+					where : { requestSessionId },
 				});
 
-				if ( newDayModuleIds.length > 0 ) {
-					await this.sessionDayModule.createMany({
-						data : newDayModuleIds.map(( dayModuleId ) => ({
-							dayModuleId,
-							requestSessionId: requestSession.id
-						})),
+                if ( currentSessionDayModules.length === 0 ) {
+                    await this.sessionDayModule.createMany({
+                        data : dayModulesId.map(( dayModuleId ) => ({
+                            dayModuleId,
+                            requestSessionId
+                        })),
+                    });
+
+                    continue;
+                }
+
+				const currentDayModuleIds	= currentSessionDayModules.map(( dm ) => dm.dayModuleId ).sort();
+				const newDayModuleIds		= dayModulesId.sort();
+				const areEqual				= currentDayModuleIds.length === newDayModuleIds.length &&
+					currentDayModuleIds.every(( id, index ) => id === newDayModuleIds[index] );
+
+				if ( !areEqual ) {
+					await this.sessionDayModule.deleteMany({
+						where : { requestSessionId },
 					});
+
+					if ( newDayModuleIds.length > 0 ) {
+						await this.sessionDayModule.createMany({
+							data : newDayModuleIds.map(( dayModuleId ) => ({
+								dayModuleId,
+								requestSessionId
+							})),
+						});
+					}
 				}
 			}
 
-			// const requestSessionData = this.#requestSessionMap( requestSession );
-			return requestSession;
+            const updatesRequestSessions = await this.requestSession.findMany({
+                where : { id: { in: updateSessionDayModulesDto.map(( dto ) => dto.requestSessionId )}},
+                select: this.#selectRequestSession
+            });
+
+            return updatesRequestSessions.map(( requestSession ) => this.#requestSessionMap( requestSession ));
 		} catch ( error ) {
-			throw PrismaException.catch( error, 'Failed to update request session' );
+			throw PrismaException.catch( error, 'Failed to update session day modules' );
 		}
 	}
